+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+# Unreleased
+
+### Changed
+
+- Fixed issue where an unnecessary newline was being added before closing curly braces.
+
 # [0.3.8] - 2025-01-16
 
 ### Changed